Add support for inner listing of addresses in event log (#243)
This requires that sub-types of multisig (anything types that are used
in fields with an inner tag) implement a trait, so I added a tag for
Multisig arguments that defaults to empty set instead if the trait
is not implemented. Some types would never implement it anyway.

I re-enabled the test that was disabled, and added a new test for
a level 2 recursion. This should be enough to tell us it covers ALL
recursive event logs.