fix: add Application::check_tx to ensure txs are validated
This is the CHANGELOG item for commit #1528957f3dad9113e39ca6ee6105b80288679371.

This change forces all transactions, even those who didn't come from the
MANY server, to go through at least a timestamp verification. This means
that old transactions that took too long to get into the block/mempool
will be rejected, but also that untrusted nodes cannot just submit
or replay transactions.

Verification is always done for signatures on execution of a tx, but this
also forces it when adding to the mempool.